% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockwts.R
\docType{data}
\name{blockwts}
\alias{blockwts}
\title{2020 Decennial Census block group weights}
\description{
A data.table with    blockwt, and
blockid or blockfips,
bgid or bgfips
used by doaggregate() to summarize the block group score of the
average person in the buffer, as the population weighted mean
of blockgroup scores of all the blocks in the buffer.
}
\details{
This is drawn from blocks2020 dataset

\preformatted{
  blockid is a unique integer ID 1 through total number of blocks, to be used as a more efficient indexing that the FIPS code is.
  blockfips is the 15 character Census Bureau FIPS code for each Census block
  
  bgid would be unique integer ID more efficient than bgfips but may just need to join on fips.
  bgfips from blocks2020 is the 12 character Census Bureau FIPS code for the parent blockgroup (i.e., the onecontaining the given block).
  and bgfips can be used to join these weights to a blockgroup dataset to get indicator scores.
  
  blockwt is the block's population weight, 
  calculated as the block population as a fraction of the parent blockgroup pop.
  Based on the latest decennial Census table of population count for each block.

  This table of weights can be used for calculating the weighted mean of each blockgroup score
  or a buffer where only some blocks of any given blockgroup are in the buffer.
  The sum of weights from some blocks tells you what fraction of its whole parent blockgroup's population count
  is in those blocks (the ones found inside a buffer, for example).
 }

See \url{https:\\www.epa.gov\ejscreen}
}
\seealso{
\link{make_blockwts}
}
