% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2020_clean.R
\name{census2020_clean}
\alias{census2020_clean}
\title{clean up download block data}
\usage{
census2020_clean(
  x,
  cols_to_keep = c("blockid", "blockfips", "bgfips", "lat", "lon", "ST", "pop")
)
}
\arguments{
\item{x}{data from census2020_read()}

\item{cols_to_keep}{optional, which (renamed) columns to retain and return}
}
\value{
same as input but drops some columns, renames some columns, etc.
}
\description{
Renames variables (columns) based on census_col_names_map,
adds columns like blockid, ST and bgfips, drops columns not needed.
}
