% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_blockwts.R
\name{make_blockwts}
\alias{make_blockwts}
\title{Make blockwts, the block weights table from decennial census pop counts}
\usage{
make_blockwts(blockfips, pop)
}
\arguments{
\item{blockfips}{Census Bureau FIPS code for each Census block, as 15-character string for each block}

\item{pop}{Population count in the block, per latest decennial census.}
}
\value{
Returns a  data.table  with blockid, blockfips (key), bgfips, and blockwt columns.
}
\description{
This only has to be done once for the decennial Census, and saved as a dataset.
Using a new decennial Census table of population count for each block,
calculate for each block its weight which is its pop as a fraction of the parent blockgroup pop.
The resulting table of weights can be used for calculating the weighted mean of each blockgroup score
for a buffer where only some blocks of any given blockgroup are in the buffer.
The sum of weights from some blocks tells you what fraction of its whole parent blockgroup's population count
is in those blocks (the ones found inside a buffer, for example).
Need bgfips or bgid for join of these weights to blockgroup dataset to get indicator scores.
}
\examples{
\dontrun{
blockwts <- make_blockwts(census2020download::blocks2020$blockfips, census2020download::blocks2020$pop)
attr(blockwts, 'year') <- 2020
saveRDS(blockwts, file = file.path(getwd(),'data/blockwts.rds'))
}
}
