% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2020_clean.R
\name{census2020_clean}
\alias{census2020_clean}
\title{clean up download block data}
\usage{
census2020_clean(x)
}
\arguments{
\item{x}{data from census2020_read()}
}
\value{
same as input but drops some columns, renames some columns, etc.
}
\description{
clean up download block data
}
